Renames endpoints
