+++ chat-server/src/main/java/com/vinnik/chat/back/config/WebSocketConfig.java
@@ -17,7 +17,7 @@ public class WebSocketConfig implements WebSocketMessageBrokerConfigurer {
 
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/chat-websocket").setAllowedOrigins("*").withSockJS();
+        registry.addEndpoint("/chat-websocket/connect").setAllowedOrigins("*").withSockJS();
     }
 
 }
